requirements.txt
# =============================================================================
# Cell2Phys Requirements
# =============================================================================
# Python >= 3.10 required
# Recommended: Configure LLM server (vLLM, Ollama, or TGI) before installation

# Core Scientific Computing
# -------------------------
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.1.0
scipy>=1.7.0,<2.0.0

# Single-Cell Biology
# --------------------
scanpy>=1.8.0,<2.0.0

# Deep Learning
# -------------
torch>=1.10.0,<3.0.0

# Progress Bars
# -------------
tqdm>=4.60.0,<5.0.0

# Vector Database (ASC Engine)
# -----------------------------
faiss-cpu>=1.7.0,<2.0.0

# Visualization
# -------------
matplotlib>=3.4.0,<4.0.0

# HTTP Requests (LLM Client)
# ---------------------------  
requests>=2.26.0,<3.0.0

# LLM Framework (RAG Engine)
# ---------------------------
langchain>=0.0.100,<1.0.0
langchain-huggingface>=0.0.3

# =============================================================================
# Installation Notes:
# - LLM Server Required: This project requires a running LLM server
#   (vLLM, Ollama, TGI, or OpenAI-compatible API)
# - Configure API endpoint in src/cell2phys/config.py before running
# - For GPU acceleration: Install PyTorch with CUDA from https://pytorch.org/
# - On Windows: Some packages may require Microsoft C++ Build Tools
# =============================================================================